Add logging to a file + package external deps
Standard Java logging api:

    val log: Logger = LoggerFactory.getLogger("marcel");
    log.info("Running.")

Also removes runGame.bat because nobody is on Windows here.

The external dependency packaging is done with
[sbt-pack](https://github.com/xerial/sbt-pack). It's needed to actually
get this to run with runGame.sh/the `halite` binary.

It packages up all deps listed in build.sbt and puts them into
`target/pack/lib` as JARs, so you run MyBot with `target/pack/lib/*`
(note the star) as the classpath.